Move bft assertions into separate BFTTest class
